+++ app/controllers/api/stacks/base_controller.rb
@@ -8,10 +8,7 @@ class Api::Stacks::BaseController < ActionController::Base
     response = ClusterAuthService.new(request.headers['Authorization']&.gsub("Bearer ", "").strip)
     if response.node || response.container
       @node = response.node.nil? ? response.container.node : response.node
-      @load_balancer = nil
-      LoadBalancer.all.each do |i|
-        @load_balancer = i if i.ext_ip.include?(@node.primary_ip)
-      end
+      @load_balancer = response.load_balancer if response.load_balancer
       @container = response.container if response.container
       @deployment = response.project if response.project
       @service = response.service if response.service